Isolating CardCarousel into its own component to allow me to merge things more smoothly later
